requirements.txt
pandas
openpyxl
streamlit
numpy
python-pptx
